Refactor import paths and consolidate generation-related imports
