+++ packages/giselle-engine/src/core/flows/resolve-trigger.ts
@@ -1,12 +1,12 @@
+import { isTriggerNode } from "@giselle-sdk/data-type";
+import { githubTriggers } from "@giselle-sdk/flow";
 import {
 	type CompletedGeneration,
 	GenerationContext,
 	type GenerationContextInput,
 	type GenerationOutput,
-	isTriggerNode,
 	type QueuedGeneration,
-} from "@giselle-sdk/data-type";
-import { githubTriggers } from "@giselle-sdk/flow";
+} from "../generations";
 import { internalSetGeneration } from "../generations/internal/set-generation";
 import { resolveTrigger as resolveGitHubTrigger } from "../github/trigger-utils";
 import type { GiselleEngineContext } from "../types";
