+++ packages/giselle-engine/src/react/generations/hooks/use-node-generations.ts
@@ -1,10 +1,7 @@
-import {
-	type Generation,
-	GenerationOrigin,
-	type NodeId,
-} from "@giselle-sdk/data-type";
+import type { NodeId } from "@giselle-sdk/data-type";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import useSWR from "swr";
+import { type Generation, GenerationOrigin } from "../../../core/generations";
 import { useFeatureFlag } from "../../feature-flags";
 import { useGiselleEngine } from "../../use-giselle-engine";
 import { useGenerationRunnerSystem } from "../contexts";
